Move IOC manager logic into epics adapter
